PRO-3360: Refactor FormatAlias so that it is in its own class
